+++ py/desispec/scripts/specex.py
@@ -17,7 +17,7 @@ from ctypes.util import find_library
 from astropy.io import fits
 
 from desiutil.log import get_logger
-from desispec.psfio import run_specex
+from desispec.pybindspecex import run_specex
 
 modext = "so"
 if sys.platform == "darwin":
